requirements.txt
blinker==1.8.2
click==8.1.7
dnspython==2.6.1
Flask==3.0.3
flask-mongoengine-3==1.0.9
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mongoengine==0.28.2
packaging==24.0
pluggy==1.5.0
pymongo==4.7.2
pytest==8.2.0
Werkzeug==3.0.3
